updatebook correction

+++ middleware/sharp.js
@@ -1,25 +1,28 @@
 const sharp = require("sharp");
 const fs = require("fs");
 
+//compression du fichier et remplacement
 const compression = (req, res, next) => {
-  const input = `images/${req.file.filename}`;
-  const name = req.file.filename.split(".")[0] + ".webp";
-  const output = `images/${name}`;
+  if (req.file) {
+    const input = `images/${req.file.filename}`;
+    const name = req.file.filename.split(".")[0] + ".webp";
+    const output = `images/${name}`;
 
-  sharp(input)
-    .webp()
-    .toFile(output)
-    .then(() => {
-      fs.unlink(input, (error) => {
-        if (error) {
-          console.log(error);
-        }
-      });
-    })
-    .catch((error) => console.log(error));
+    sharp(input)
+      .webp()
+      .toFile(output)
+      .then(() => {
+        fs.unlink(input, (error) => {
+          if (error) {
+            console.log(error);
+          }
+        });
+      })
+      .catch((error) => console.log(error));
 
-  req.file.path = output;
-  req.file.filename = name;
+    req.file.path = output;
+    req.file.filename = name;
+  }
   next();
 };
